requirements.txt
solana == 0.35.0

solders == 0.21.0

openai == 0.27.0

flask == 2.2.2

python-dotenv == 0.20.0

requests == 2.28.1